requirements.txt

fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
PyJWT==2.8.0
requests==2.32.3
boto3
